+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## [Unreleased]
 
+### Added
+
+- feat: add go template functions to render a form, a field, etc.
+- feat: add constraints (isodd, iseven)
+
+### Changed
+
+- feat: replace templates with gomponents
+
 ## v1.1.0
 
 ### Added